re-factor: change jQuery link
Add jQuery link from previous exercise (to do list app) instead of the link provided by the Bootstrap page, because it was JQuery Slim Build, which is a smaller version of jQuery not including $.ajax(). This is why the modified link was not recommended by the exercise.
Also, change the number in the jQuery link from 3.5.1. to 3.3.1 to match the number in the link provided by the bootstrap page https://getbootstrap.com/docs/4.3/getting-started/introduction/.